CS Go Friends

This function will return all the CS Go friends of the user_id (input).

csgo_api_friend(api_key, user_id)

Arguments

api_key
string with the key provided by the steam API.

PS: If you don`t have a API key yet run vignette("auth", package = "CSGo") and follow the presented steps.

user_id
string with the steam user ID.

Steam ID is the NUMBER OR NAME at the end of your steam profile URL. ex: '76561198263364899'.

PS: The user should have a public status.

Value

data frame with all the CS Go friends of the user ID.

Examples

if (FALSE) {
## It is necessary to fill the "api_key" parameter to run the example

df_friend <- csgo_api_friend(api_key = 'XXX', user_id = '76561198263364899')
}